Allow for different record directory, rather than just ~/wth
wth.sh now assumes that records are either in $WTHDIR, $XDG_DATA_DIR/wth
or ~/.local/share/wth as a fallback. This change was prompted by the
annoyingness of having different programs (including this one) clobber
up my $HOME, and the discovery of the $XDG_DATA_DIR standard [1], which
is supposed to solve this.

1: https://specifications.freedesktop.org/basedir-spec/latest/